@@ -1,38 +1,31 @@
 {
-  "name": "minube/geo-timezone",
-  "homepage": "http://minube.com",
-  "description": "PHP library for timezone calculation from a particular location and date.",
-  "license": "MIT",
-  "repositories": [
-    {
-      "type": "git",
-      "url": "https://github.com/phayes/geoPHP",
-      "autoload": {
-        "files": [
-          "vendor/phayes/geophp/geoPHP.inc"
-        ]
-      }
+    "name": "tochka-developers/geo-timezone",
+    "homepage": "http://minube.com",
+    "description": "PHP library for timezone calculation from a particular location and date.",
+    "license": "MIT",
+    "require": {
+        "php": "8.2.*|8.3.*",
+        "ext-dom": "*",
+        "ext-mbstring": "*",
+        "ext-zip": "*",
+        "ext-xml": "*",
+        "tochka-developers/geophp": "^0.1",
+        "psr/log": "^3.0"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^10.0",
+        "guzzlehttp/guzzle": "^6.3",
+        "laravel/pint": "^1.14",
+        "vimeo/psalm": "^5.23"
+    },
+    "autoload": {
+        "psr-4": {
+            "Tochka\\GeoTimeZone\\": "src/",
+            "Tests\\": "tests/"
+        }
+    },
+    "scripts": {
+        "pint": "vendor/bin/pint",
+        "psalm": "vendor/bin/psalm --output-format=phpstorm --no-cache"
     }
-  ],
-  "require": {
-    "php": ">=7.0",
-    "ext-dom": ">=7.0",
-    "ext-mbstring": ">=7.0",
-    "ext-zip": "*",
-    "ext-xml":">=7.0",
-    "phayes/geoPHP": "1.2.*"
-  },
-  "require-dev": {
-    "phpunit/phpunit": "5.7.4",
-    "guzzlehttp/guzzle": "6.3.0"
-  },
-  "autoload": {
-    "psr-4": {
-      "GeoTimeZone\\": "src/GeoTimeZone/",
-      "Tests\\": "tests/"
-    }
-  },
-  "config": {
-    "vendor-dir": "vendor"
-  }
 }
